weather-forecast/requirements.txt
dvc[gdrive]
tensorflow
scikit-learn
pandas
numpy
requests
streamlit
plotly
pyyaml
